+++ ulox/ulox.core.bench/BenchmarkScripts.cs
@@ -59,4 +59,86 @@ match i
 0: i = 3;
 }");
     }
+
+    public static class BouncingBallProfileScript
+    {
+		public static readonly Script Script = new Script(nameof(BouncingBallProfileScript), @"
+var dt = 0;
+var limit = 5;
+var numBallsToSpawn = 200;
+
+var ballName = ""BouncyBall"";
+var balls = [];
+
+fun RandVec2()
+{
+	return (-3,3);
+}
+
+class Ball
+{
+	init()
+	{
+		var (x,y) = RandVec2();
+		this.posx = x;
+		this.posy = y;
+
+		var (vx,vy) = RandVec2();
+		this.velx = vx;
+		this.vely = vy;
+		
+		//this.go = CreateFromPrefab(ballName);
+	}
+
+	Tick()
+	{
+  		this.posx = this.posx + this.velx * dt;
+  		this.posy = this.posy + this.vely * dt;
+
+		this.Contain(limit);
+		//this.UpdateGoPosition();
+	}
+	
+	Contain(lim)
+	{
+		var x = this.posx;
+		var y = this.posy;
+		var vx = this.velx;
+		var vy = this.vely;
+
+		if((x < -lim and vx < 0) or 
+		  (x > lim  and vx > 0) )
+		{
+			vx *= -1;
+			this.velx = vx;
+		}
+
+		if((y < -lim and vy < 0) or
+		  (y > lim  and vy > 0) )
+		{
+			vy *= -1;
+			this.vely = vy;
+		}
+	}
+}
+
+fun SetupGame()
+{
+	print (""Setting Up Game"");
+
+	for(var i = 0; i < numBallsToSpawn; i += 1)
+	{
+		balls.Add(Ball());
+	}
+}
+
+fun Update()
+{
+	print (""Updating"");
+	loop(balls)
+	{
+		item.Tick();
+	}
+}");
+    }
 }
